+++ rollingstock/SteamLocomotive.java
@@ -10,13 +10,13 @@ package rollingstock;
 public class SteamLocomotive extends Locomotive
 {
     // instance variables - replace the example below with your own
-    private float pascalPower;
+    private double pascalPower;
 
     /**
      * Constructor for objects of class SteamLocomotive
      */
     
-     public SteamLocomotive(int trackGauge ,int KW, float pascalPower)
+     public SteamLocomotive(int trackGauge ,int KW, double pascalPower)
     {
         // initialise instance variables
         super(trackGauge, KW); 
@@ -28,7 +28,7 @@ public class SteamLocomotive extends Locomotive
      * @param  y  a sample parameter for a method
      * @return    the sum of x and y
      */
-    public float getPower()
+    public double getPower()
     {
         // put your code here
         return  this.pascalPower;
